requirements.txt
speechrecognition>=3.10.0
pyaudio>=0.2.14
gTTS>=2.5.1
playsound3>=3.3.0
requests>=2.31.0
python-dotenv>=1.0.1